+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 *   Used new joulescope.bootloaders_run_application() for better error handling.
 *   Improved logging to include header with joulescope and platform information.
 *   Added "--file_log_level" command line option.
+*   Display Qt Window when configuration initialization fails.
 
 
 ## 0.4.2